Re-ordered flow sequence for clarity
This is the 'Short' version of the App Connect Integrated Car Crash Insurance flow. It analyses the submitted photo and checks it has a car in it, creates a case in SalesForce, adds tone analysis as a comment to the case and adds the photograph as an attachment. This version does *not* include the Language Translation and Integration with ServiceNow components.